+++ index.js
@@ -3,7 +3,8 @@
 const ServiceInstancesMonitor = require('./src/ServiceInstancesMonitor');
 const ServiceInstance         = require('./src/ServiceInstance');
 const ServiceInstances        = require('./src/ServiceInstances');
-const ServiceInstanceStatus   = require('./src/ServiceInstanceStatus');
+const ServiceInstanceInfo     = require('./src/ServiceInstanceInfo');
+const builders                = require('./src/Factory');
 
 const Errors = require('./src/Error');
 
@@ -11,6 +12,7 @@ module.exports = {
     ServiceInstancesMonitor,
     ServiceInstance,
     ServiceInstances,
-    ServiceInstanceStatus,
-    Errors
+    ServiceInstanceInfo,
+    Errors,
+    builders
 };