+++ tests/Functional/ServiceInstancesMonitor.test.js
@@ -5,17 +5,17 @@ const {assertThrowsAsync} = require('../support/helpers');
 const consul = require('consul');
 const nock = require('nock');
 const assert = require('chai').assert;
+const sinon = require('sinon');
 const deepFreeze = require('deep-freeze');
 const getPort = require('get-port');
 const ServiceInstance = require('src/ServiceInstance');
-const ServiceInstanceStatus = require('src/ServiceInstanceStatus');
 const ServiceInstances = require('src/ServiceInstances');
 const ServiceInstancesMonitor = require('src/ServiceInstancesMonitor');
 const {WatchError, WatchTimeoutError, InvalidDataError} = require('src/Error');
 
 const nockTestParams = require('./nock.data');
 
-describe('ServiceInstancesMonitor::constructor', function () {
+describe('ServiceInstancesMonitor methods tests', function () {
 
     const consulHost = '127.0.0.1';
     let consulPort;
@@ -55,8 +55,6 @@ describe('ServiceInstancesMonitor::constructor', function () {
         assert.isFalse(monitor.isWatchHealthy());
         assert.instanceOf(monitor.getInstances(), ServiceInstances);
         assert.isEmpty(monitor.getInstances().getHealthy());
-        assert.isEmpty(monitor.getInstances().getOnMaintenance());
-        assert.isEmpty(monitor.getInstances().getOverloaded());
         assert.isEmpty(monitor.getInstances().getUnhealthy());
     });
 
@@ -71,8 +69,6 @@ describe('ServiceInstancesMonitor::constructor', function () {
         assert.isFalse(monitor.isWatchHealthy());
         assert.instanceOf(monitor.getInstances(), ServiceInstances);
         assert.isEmpty(monitor.getInstances().getHealthy());
-        assert.isEmpty(monitor.getInstances().getOnMaintenance());
-        assert.isEmpty(monitor.getInstances().getOverloaded());
         assert.isEmpty(monitor.getInstances().getUnhealthy());
         assert.strictEqual(returnedValue, monitor);
     });
@@ -80,9 +76,11 @@ describe('ServiceInstancesMonitor::constructor', function () {
     it('start monitor fails if port is closed', async function () {
         const monitor = new ServiceInstancesMonitor(options, consulClient);
 
-        assertThrowsAsync(async () => {
-            await monitor.startService();
-        }, WatchError, /connect ECONNREFUSED/);
+        await assertThrowsAsync(
+            () => monitor.startService(),
+            WatchError,
+            /connect ECONNREFUSED/
+        );
     });
 
     it('start monitor fails due to consul response timeout - no requests after timeout', async function () {
@@ -90,7 +88,6 @@ describe('ServiceInstancesMonitor::constructor', function () {
         // then after extra options.timeoutMsec time response from nock must be returned
         // and monitor must ignore that update
 
-        this.slow(options.timeoutMsec * 8);
         this.timeout(options.timeoutMsec * 4);
 
         const nockInstance = nock(consulHostAndPort)
@@ -106,9 +103,11 @@ describe('ServiceInstancesMonitor::constructor', function () {
             changeFired = true;
         });
 
-        assertThrowsAsync(async () => {
-            await monitor.startService();
-        }, WatchTimeoutError, 'Initial consul watch request was timed out');
+        await assertThrowsAsync(
+            () => monitor.startService(),
+            WatchTimeoutError,
+            'Initial consul watch request was timed out'
+        );
 
         const waitFn = () => {
             return new Promise(resolve => {
@@ -123,12 +122,16 @@ describe('ServiceInstancesMonitor::constructor', function () {
         assert.isFalse(monitor.isInitialized());
         assert.isFalse(monitor.isWatchHealthy());
         assert.isEmpty(monitor.getInstances().getHealthy());
-        assert.isEmpty(monitor.getInstances().getOnMaintenance());
-        assert.isEmpty(monitor.getInstances().getOverloaded());
         assert.isEmpty(monitor.getInstances().getUnhealthy());
     });
 
     it('monitor becomes initialized and watch becomes healthy after start of monitor', async function () {
+        const expectedConsulHeaders = {
+            'x-consul-index': nockTestParams.firstResponseHeaders['X-Consul-Index'],
+            'x-consul-knownleader': nockTestParams.firstResponseHeaders['X-Consul-Knownleader'],
+            'x-consul-lastcontact': nockTestParams.firstResponseHeaders['X-Consul-Lastcontact']
+        };
+
         const firstRequestIndex = 0;
         // blocking queries read X-Consul-Index header and make next request using that value as index
         const secondRequestIndex = nockTestParams.firstResponseHeaders['X-Consul-Index'];
@@ -145,9 +148,17 @@ describe('ServiceInstancesMonitor::constructor', function () {
 
         assert.isTrue(monitor.isInitialized());
         assert.isTrue(monitor.isWatchHealthy());
+        assert.deepEqual(monitor.getConsulHeaders(), expectedConsulHeaders);
+        monitor.stopService();
     });
 
     it('check of initial list of nodes received from startService', async function () {
+        const expectedConsulHeaders = {
+            'x-consul-index': nockTestParams.firstResponseHeaders['X-Consul-Index'],
+            'x-consul-knownleader': nockTestParams.firstResponseHeaders['X-Consul-Knownleader'],
+            'x-consul-lastcontact': nockTestParams.firstResponseHeaders['X-Consul-Lastcontact']
+        };
+
         const expectedNode1 = new ServiceInstance(
             nockTestParams.firstResponseBody[0].Node.TaggedAddresses.lan,
             nockTestParams.firstResponseBody[0].Node.TaggedAddresses.wan,
@@ -156,15 +167,7 @@ describe('ServiceInstancesMonitor::constructor', function () {
             nockTestParams.firstResponseBody[0].Node.Node,
             nockTestParams.firstResponseBody[0].Service.ID,
             nockTestParams.firstResponseBody[0].Service.Tags,
-            new ServiceInstanceStatus(
-                nockTestParams.loadData1.pid,
-                nockTestParams.loadData1.status,
-                nockTestParams.loadData1.mem.total,
-                nockTestParams.loadData1.mem.free,
-                nockTestParams.loadData1.cpu.usage,
-                nockTestParams.loadData1.cpu.count,
-                nockTestParams.loadData1
-            )
+            null
         );
 
         const expectedNode2 = new ServiceInstance(
@@ -175,15 +178,7 @@ describe('ServiceInstancesMonitor::constructor', function () {
             nockTestParams.firstResponseBody[1].Node.Node,
             nockTestParams.firstResponseBody[1].Service.ID,
             nockTestParams.firstResponseBody[1].Service.Tags,
-            new ServiceInstanceStatus(
-                nockTestParams.loadData1.pid,
-                nockTestParams.loadData1.status,
-                nockTestParams.loadData1.mem.total,
-                nockTestParams.loadData1.mem.free,
-                nockTestParams.loadData1.cpu.usage,
-                nockTestParams.loadData1.cpu.count,
-                nockTestParams.loadData1
-            )
+            null
         );
 
         const firstRequestIndex = 0;
@@ -199,11 +194,10 @@ describe('ServiceInstancesMonitor::constructor', function () {
 
         const monitor = new ServiceInstancesMonitor(options, consulClient);
         const initialInstances = await monitor.startService();
+        monitor.stopService();
 
         assert.instanceOf(initialInstances, ServiceInstances);
         assert.lengthOf(initialInstances.getHealthy(), 2);
-        assert.isEmpty(initialInstances.getOnMaintenance());
-        assert.isEmpty(initialInstances.getOverloaded());
         assert.isEmpty(initialInstances.getUnhealthy());
 
         const [node1, node2] = initialInstances.getHealthy();
@@ -212,6 +206,7 @@ describe('ServiceInstancesMonitor::constructor', function () {
         assert.instanceOf(node2, ServiceInstance);
         assert.deepEqual(node1, expectedNode1);
         assert.deepEqual(node2, expectedNode2);
+        assert.deepEqual(monitor.getConsulHeaders(), expectedConsulHeaders);
     });
 
     it('initial list of nodes is the same as received from getter', async function () {
@@ -229,6 +224,7 @@ describe('ServiceInstancesMonitor::constructor', function () {
         const monitor = new ServiceInstancesMonitor(options, consulClient);
         const initialInstances = await monitor.startService();
         const instancesFromGetter = monitor.getInstances();
+        monitor.stopService();
 
         assert.strictEqual(initialInstances, instancesFromGetter);
     });
@@ -266,7 +262,7 @@ describe('ServiceInstancesMonitor::constructor', function () {
 
         const waitFn = () => {
             return new Promise(resolve => {
-                setTimeout(resolve, 0);
+                setTimeout(resolve, 10);
             });
         };
 
@@ -295,15 +291,12 @@ describe('ServiceInstancesMonitor::constructor', function () {
 
         assert.instanceOf(initialInstances, ServiceInstances);
         assert.isEmpty(initialInstances.getHealthy());
-        assert.isEmpty(initialInstances.getOnMaintenance());
-        assert.isEmpty(initialInstances.getOverloaded());
         assert.isEmpty(initialInstances.getUnhealthy());
+
+        monitor.stopService();
     });
 
     it('reaction on 500 error from consul during startService', async function () {
-        this.slow(options.timeoutMsec * 5);
-        this.timeout(options.timeoutMsec * 5);
-
         const firstRequestIndex = 0;
 
         const nockInstance = nock(consulHostAndPort)
@@ -319,9 +312,7 @@ describe('ServiceInstancesMonitor::constructor', function () {
             changeFired = true;
         });
 
-        assertThrowsAsync(async () => {
-            await monitor.startService();
-        }, WatchError, 'internal server error');
+        await assertThrowsAsync(() => monitor.startService(), WatchError, 'internal server error');
 
 
         const waitFn = () => {
@@ -337,15 +328,12 @@ describe('ServiceInstancesMonitor::constructor', function () {
         assert.isFalse(monitor.isInitialized());
         assert.isFalse(monitor.isWatchHealthy());
         assert.isEmpty(monitor.getInstances().getHealthy());
-        assert.isEmpty(monitor.getInstances().getOnMaintenance());
-        assert.isEmpty(monitor.getInstances().getOverloaded());
         assert.isEmpty(monitor.getInstances().getUnhealthy());
+        assert.isFalse(monitor._isWatcherRegistered());
+        assert.deepEqual(monitor.getConsulHeaders(), {});
     });
 
     it('reaction on 400 error from consul during startService', async function () {
-        this.slow(options.timeoutMsec * 15);
-        this.timeout(options.timeoutMsec * 5);
-
         const firstRequestIndex = 0;
 
         const nockInstance = nock(consulHostAndPort)
@@ -361,9 +349,7 @@ describe('ServiceInstancesMonitor::constructor', function () {
             changeFired = true;
         });
 
-        assertThrowsAsync(async () => {
-            await monitor.startService();
-        }, WatchError, 'bad request');
+        await assertThrowsAsync(() => monitor.startService(), WatchError, 'bad request');
 
 
         const waitFn = () => {
@@ -379,30 +365,17 @@ describe('ServiceInstancesMonitor::constructor', function () {
         assert.isFalse(monitor.isInitialized());
         assert.isFalse(monitor.isWatchHealthy());
         assert.isEmpty(monitor.getInstances().getHealthy());
-        assert.isEmpty(monitor.getInstances().getOnMaintenance());
-        assert.isEmpty(monitor.getInstances().getOverloaded());
         assert.isEmpty(monitor.getInstances().getUnhealthy());
+        assert.isFalse(monitor._isWatcherRegistered());
+        assert.deepEqual(monitor.getConsulHeaders(), {});
     });
 
     it('emission of error on initial data', async function () {
-        const expectedNode2 = new ServiceInstance(
-            nockTestParams.firstResponseBody[1].Node.TaggedAddresses.lan,
-            nockTestParams.firstResponseBody[1].Node.TaggedAddresses.wan,
-            nockTestParams.firstResponseBody[1].Service.Port,
-            nockTestParams.firstResponseBody[1].Node.Address,
-            nockTestParams.firstResponseBody[1].Node.Node,
-            nockTestParams.firstResponseBody[1].Service.ID,
-            nockTestParams.firstResponseBody[1].Service.Tags,
-            new ServiceInstanceStatus(
-                nockTestParams.loadData1.pid,
-                nockTestParams.loadData1.status,
-                nockTestParams.loadData1.mem.total,
-                nockTestParams.loadData1.mem.free,
-                nockTestParams.loadData1.cpu.usage,
-                nockTestParams.loadData1.cpu.count,
-                nockTestParams.loadData1
-            )
-        );
+        const expectedConsulHeaders = {
+            'x-consul-index': nockTestParams.firstResponseHeaders['X-Consul-Index'],
+            'x-consul-knownleader': nockTestParams.firstResponseHeaders['X-Consul-Knownleader'],
+            'x-consul-lastcontact': nockTestParams.firstResponseHeaders['X-Consul-Lastcontact']
+        };
 
         const firstResponseBody = _.cloneDeep(nockTestParams.firstResponseBody);
         firstResponseBody[0].Checks[1].Name = 'Name of check that will not match checkNameWithStatus';
@@ -425,7 +398,7 @@ describe('ServiceInstancesMonitor::constructor', function () {
 
         const waitFn = () => {
             return new Promise(resolve => {
-                setTimeout(resolve, 0);
+                setTimeout(resolve, 100);
             });
         };
 
@@ -438,17 +411,9 @@ describe('ServiceInstancesMonitor::constructor', function () {
         const initialInstances = await monitor.startService();
 
         assert.lengthOf(errors, 0);
-
         assert.instanceOf(initialInstances, ServiceInstances);
         assert.lengthOf(initialInstances.getHealthy(), 1);
-        assert.isEmpty(initialInstances.getOnMaintenance());
-        assert.isEmpty(initialInstances.getOverloaded());
         assert.isEmpty(initialInstances.getUnhealthy());
-
-        const [node2] = initialInstances.getHealthy();
-
-        assert.instanceOf(node2, ServiceInstance);
-        assert.deepEqual(node2, expectedNode2);
 
         await waitFn();
 
@@ -456,5 +421,162 @@ describe('ServiceInstancesMonitor::constructor', function () {
         assert.instanceOf(errors[0], expectedErrorType);
         assert.strictEqual(errors[0].message, expectedErrorMessage);
         assert.deepEqual(errors[0].extra, expectedErrorExtra);
+        assert.deepEqual(monitor.getConsulHeaders(), expectedConsulHeaders);
+        monitor.stopService();
+    });
+
+    it('auto restart service on watcher "end" (response with status 400)', async function () {
+        const expectedConsulHeaders = {
+            'x-consul-index': nockTestParams.firstResponseHeaders['X-Consul-Index'],
+            'x-consul-knownleader': nockTestParams.firstResponseHeaders['X-Consul-Knownleader'],
+            'x-consul-lastcontact': nockTestParams.firstResponseHeaders['X-Consul-Lastcontact']
+        };
+
+        const firstRequestIndex = 0;
+        const secondRequestIndex = nockTestParams.firstResponseHeaders['X-Consul-Index'];
+        const secondResponseBody = [nockTestParams.firstResponseBody[0]];
+
+        const nockInstance = nock(consulHostAndPort)
+            .get(`/v1/health/service/${options.serviceName}`).query({index: firstRequestIndex, wait: '60s'})
+            .reply(200, nockTestParams.firstResponseBody, nockTestParams.firstResponseHeaders)
+            .get(`/v1/health/service/${options.serviceName}`).query({index: secondRequestIndex, wait: '60s'})
+            .reply(400, 'Not available')
+            .get(`/v1/health/service/${options.serviceName}`).query({index: firstRequestIndex, wait: '60s'})
+            .reply(200, secondResponseBody, nockTestParams.firstResponseHeaders)
+            .get(`/v1/health/service/${options.serviceName}`).query({index: secondRequestIndex, wait: '60s'})
+            .delay(60000)
+            .reply(400, 'Not available');
+
+        let changeFiredCount = 0;
+        let firedInstances;
+        let healthyFiredCount = 0;
+        let unhealthyFiredCount = 0;
+        const errors = [];
+        const monitor = new ServiceInstancesMonitor(options, consulClient);
+
+        sinon.spy(monitor, '_retryStartService');
+
+        monitor.on('changed', instances => {
+            changeFiredCount++;
+            firedInstances = instances;
+        });
+
+        monitor.on('error', err => {
+            errors.push(err);
+        });
+
+        monitor.on('unhealthy', () => {
+            unhealthyFiredCount++;
+        });
+
+        monitor.on('healthy', () => {
+            healthyFiredCount++;
+        });
+
+        await monitor.startService();
+
+        const waitFn = () => {
+            return new Promise(resolve => {
+                setTimeout(resolve, options.timeoutMsec / 2);
+            });
+        };
+
+        await waitFn();
+
+        assert.isTrue(nockInstance.isDone());
+        assert.equal(changeFiredCount, 1);
+        assert.equal(unhealthyFiredCount, 1);
+        assert.equal(healthyFiredCount, 1);
+        assert.isTrue(monitor.isInitialized());
+        assert.isTrue(monitor.isWatchHealthy());
+        assert.deepEqual(firedInstances, monitor.getInstances());
+        assert.lengthOf(monitor.getInstances().getHealthy(), secondResponseBody.length);
+        assert.isEmpty(monitor.getInstances().getUnhealthy());
+        assert.lengthOf(errors, 1);
+        assert.instanceOf(errors[0], WatchError);
+        assert.isTrue(monitor._isWatcherRegistered());
+        assert.isTrue(monitor._retryStartService.calledOnce);
+        assert.isTrue(monitor._retryStartService.calledWithExactly());
+        assert.deepEqual(monitor.getConsulHeaders(), expectedConsulHeaders);
+        monitor.stopService();
+    });
+
+    it('service goes to "unhealthy" state on response with status 500 and ' +
+        'returns to "healthy" after success response', async function () {
+        const firstRequestIndex = 0;
+        const secondRequestIndex = nockTestParams.firstResponseHeaders['X-Consul-Index'];
+        const secondResponseBody = [nockTestParams.firstResponseBody[0]];
+        const secondResponseHeaders = _.cloneDeep(nockTestParams.firstResponseHeaders);
+        secondResponseHeaders['X-Consul-Index'] += 1;
+        const thirdRequestIndex = secondResponseHeaders['X-Consul-Index'];
+
+        const expectedConsulHeaders = {
+            'x-consul-index': secondResponseHeaders['X-Consul-Index'],
+            'x-consul-knownleader': secondResponseHeaders['X-Consul-Knownleader'],
+            'x-consul-lastcontact': secondResponseHeaders['X-Consul-Lastcontact']
+        };
+
+        const nockInstance = nock(consulHostAndPort)
+            .get(`/v1/health/service/${options.serviceName}`).query({index: firstRequestIndex, wait: '60s'})
+            .reply(200, nockTestParams.firstResponseBody, nockTestParams.firstResponseHeaders)
+            .get(`/v1/health/service/${options.serviceName}`).query({index: secondRequestIndex, wait: '60s'})
+            .reply(500, 'Internal error')
+            .get(`/v1/health/service/${options.serviceName}`).query({index: secondRequestIndex, wait: '60s'})
+            .reply(200, secondResponseBody, secondResponseHeaders)
+            .get(`/v1/health/service/${options.serviceName}`).query({index: thirdRequestIndex, wait: '60s'})
+            .delay(60000)
+            .reply(400, 'Not available');
+
+        let changeFiredCount = 0;
+        let firedInstances;
+        let healthyFiredCount = 0;
+        let unhealthyFiredCount = 0;
+        const errors = [];
+        const monitor = new ServiceInstancesMonitor(options, consulClient);
+
+        sinon.spy(monitor, '_retryStartService');
+
+        monitor.on('changed', instances => {
+            changeFiredCount++;
+            firedInstances = instances;
+        });
+
+        monitor.on('error', err => {
+            errors.push(err);
+        });
+
+        monitor.on('unhealthy', () => {
+            unhealthyFiredCount++;
+        });
+
+        monitor.on('healthy', () => {
+            healthyFiredCount++;
+        });
+
+        await monitor.startService();
+
+        const waitFn = () => {
+            return new Promise(resolve => {
+                setTimeout(resolve, options.timeoutMsec / 2);
+            });
+        };
+
+        await waitFn();
+
+        assert.isTrue(nockInstance.isDone());
+        assert.equal(changeFiredCount, 1);
+        assert.equal(unhealthyFiredCount, 1);
+        assert.equal(healthyFiredCount, 1);
+        assert.isTrue(monitor.isInitialized());
+        assert.isTrue(monitor.isWatchHealthy());
+        assert.deepEqual(firedInstances, monitor.getInstances());
+        assert.lengthOf(monitor.getInstances().getHealthy(), secondResponseBody.length);
+        assert.isEmpty(monitor.getInstances().getUnhealthy());
+        assert.lengthOf(errors, 1);
+        assert.instanceOf(errors[0], WatchError);
+        assert.isTrue(monitor._isWatcherRegistered());
+        assert.isTrue(monitor._retryStartService.notCalled);
+        assert.deepEqual(monitor.getConsulHeaders(), expectedConsulHeaders);
+        monitor.stopService();
     });
 });