+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 ## Changelog
 
+### 2.0.0
+
+UPDATES:
+
+- [#20553](https://redmine.hwtool.net/issues/20553) Add forwarding headers X-Consul-*
+- [#20559](https://redmine.hwtool.net/issues/20559) Add method "getAll()" to ServiceInstances class
+- [#20521](https://redmine.hwtool.net/issues/20521) Fix bug with default timeout for Consul Watch
+- [#20123](https://redmine.hwtool.net/issues/20123) Add to "consul-service-health-monitor" module functionality of an auto-reconnect to Сonsul
+- [#19681](https://redmine.hwtool.net/issues/19681) Extending consul-service-health-monitor to extract custom data from output
+- support auto-reconnect to Consul 
+- removed `emergencyStop` event from `ServiceInstancesMonitor`
+- removed `ServiceInstanceStatus` class and added `ServiceInstanceInfo`
+- removed methods `addOnMaintenance`, `addOverloaded`, `getOverloaded` and `getOnMaintenance` from `ServiceInstances` class.
+- removed method `getStatus` from `ServiceInstance` and added `getInfo`
+- added `extractors` argument to `ServiceInstancesMonitor::constructor`
+- added `healthy` and `unhealthy` events to `ServiceInstancesMonitor`
+
 ### 1.4.0
 
 - [#20405](https://redmine.hwtool.net/issues/20405) Implement fallback to healthy state after watch error.